requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.6.0
apprise==0.9.5.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bidict==0.22.1
cachetools==4.2.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cheroot==10.0.0
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
coveralls==1.11.1
cryptography==41.0.7
dateparser==1.2.0
defusedxml==0.7.1
distlib==0.3.7
dj-rest-auth==5.0.1
Django==4.2.9
django-admin-volt==1.0.10
django-allauth==0.57.0
django-appconf==1.0.6
django-compressor==4.4
django-cors-headers==3.2.1
django-crispy-forms==2.1
django-htmx==1.17.2
django-ninja==1.1.0
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.7.1
dnspython==1.16.0
docopt==0.6.2
drf-yasg==1.21.4
eventlet==0.30.2
Faker==22.1.0
filelock==3.13.1
flake8==3.8.4
Flask==2.1.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.10
Flask-SocketIO==5.0.1
frozenlist==1.4.0
greenlet==3.0.1
gunicorn==21.2.0
h11==0.14.0
httpie==3.2.2
idna==3.6
importlib-metadata==6.9.0
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
jaraco.functools==4.0.0
Jinja2==3.1.2
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mccabe==0.6.1
mdurl==0.1.2
more-itertools==10.1.0
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
pathlib==1.0.1
platformdirs==4.1.0
psutil==5.9.6
psycopg2-binary==2.9.9
pycodestyle==2.6.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==2.2.0
Pygments==2.17.2
PyJWT==2.8.0
PySocks==1.7.1
python-binance==1.0.12
python-dateutil==2.8.2
python-decouple==3.0
python-dotenv==0.12.0
python-engineio==4.8.0
python-socketio==5.2.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
rcssmin==1.1.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
responses==0.12.1
rich==13.7.0
rjsmin==1.2.1
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
schedule==1.1.0
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==1.4.15
sqlitedict==1.7.0
sqlparse==0.4.4
static3==0.7.0
toposort==1.10
typing_extensions==4.8.0
tzlocal==5.2
ujson==5.8.0
unicorn-binance-websocket-api==1.34.2
unicorn-fy==0.11.0
Unipath==1.0
unittest-xml-reporting==3.0.4
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.25.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==10.0
Werkzeug==3.0.1
whitenoise==6.5.0
wsproto==1.2.0
yarl==1.9.3
zipp==3.17.0
